move TextureRegion and TextureSlice into RefreshTypes
